backend/src/core/security.py
from datetime import datetime, timedelta
from typing import Any, Union
from jose import jwt
import bcrypt
import hashlib
from src.core.config import settings

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Verify a password against a hash.
    The password is first hashed with SHA256 to ensure it fits within bcrypt's 72-byte limit.
    """
    # 1. SHA256 pre-hash to handle arbitrary length
    hashed_input = hashlib.sha256(plain_password.encode('utf-8')).hexdigest()
    
    # 2. Verify with bcrypt
    # bcrypt.checkpw requires bytes for both arguments
    return bcrypt.checkpw(
        hashed_input.encode('utf-8'), 
        hashed_password.encode('utf-8')
    )

def get_password_hash(password: str) -> str:
    """
    Hash a password using bcrypt.
    The password is first hashed with SHA256 to ensure it fits within bcrypt's 72-byte limit.
    """
    # 1. SHA256 pre-hash
    hashed_input = hashlib.sha256(password.encode('utf-8')).hexdigest()
    
    # 2. Hash with bcrypt
    # bcrypt.hashpw requires bytes, returns bytes
    salt = bcrypt.gensalt()
    hashed = bcrypt.hashpw(hashed_input.encode('utf-8'), salt)
    
    # Return as string for database storage
    return hashed.decode('utf-8')

def create_access_token(subject: Union[str, Any], expires_delta: timedelta = None) -> str:
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    
    to_encode = {"exp": expire, "sub": str(subject)}
    encoded_jwt = jwt.encode(to_encode, settings.SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt
